fix: close dev-SRS verifier gap in readiness gate + review cleanup
Remediation from a full code review. The committed on-chain verifier
was a DEV-SRS (forgeable) artifact that the readiness gate could not
detect; this makes the gate catch it, plus fmt/clippy/cleanup.

Security — verifier provenance (tools/src/readiness.rs):
- New check 1e: scan the self-describing banner that
  claim_evm_roundtrip.rs stamps into every emitted verifier. A DEV-SRS
  banner is a HARD regression (forgeable + won't agree with the pinned
  production SRS); PINNED-SRS passes. Closes the gap where the gate was
  green on a dev-SRS verifier.
- New check 1f: byte-pin the canonical verifier via
  MAINNET_VERIFIER_SHA256 (empty => known blocker until the operator
  blesses the pinned-SRS artifact; mismatch => hard regression).

Cleanup:
- Gate test_generate_claim_solidity_verifier's verifier write behind
  ZKMIST_EMIT_VERIFIER so `cargo test` no longer silently overwrites the
  committed verifier with a dev-SRS one on every CI run.
- Delete orphaned duplicate contracts/Halo2Verifier.axiom.sol (outside
  foundry src/, never compiled, divergent pragma).
- Delete broken untracked fix_tests.py (regex referenced nonexistent
  group \4).
- Fix stale `num_instance = 3` / `[root, nullifier, recipient]` comments
  -> 4 instances incl. chain_id.

Style/lint:
- cargo fmt --all.
- Fix 8 clippy lints: unused imports (gen_evm_verifier_shplonk,
  ff::PrimeField), unnecessary u64->u64 cast, 3x `% 2 == 0` ->
  is_multiple_of, doc overindent, and convert verify_srs's
  constant-condition assert to a compile-time `const { assert!(..) }`
  guard.

+++ cli/src/halo2_prover_axiom.rs
@@ -31,7 +31,7 @@ use halo2_base::{
     utils::{fs::gen_srs, modulus},
 };
 use num_bigint::BigUint;
-use snark_verifier_sdk::evm::{gen_evm_proof_shplonk, gen_evm_verifier_shplonk, gen_evm_verifier_sol_code};
+use snark_verifier_sdk::evm::{gen_evm_proof_shplonk, gen_evm_verifier_sol_code};
 use snark_verifier_sdk::SHPLONK;
 
 /// Cache dir for the KZG SRS (~/.zkmist/cache).
@@ -165,7 +165,7 @@ pub fn generate_v2_proof_axiom(
     let mut recip_padded = [0u8; 32];
     recip_padded[12..32].copy_from_slice(recipient);
     let recipient_fr = bytes_be_to_fr(&recip_padded);
-    let chain_id_fr = Fr::from(crate::constants::CHAIN_ID as u64);
+    let chain_id_fr = Fr::from(crate::constants::CHAIN_ID);
 
     // Nullifier (native, halo2-base convention): poseidon(privkey mod p_BN254, domain).
     let p_bn254: BigUint = modulus::<Fr>();
@@ -261,7 +261,8 @@ pub fn generate_v2_proof_axiom(
     // Forge test against the `gen-roundtrip-fixture` output. Enable for an
     // explicit one-off check: `ZKMIST_GEN_VERIFIER=1 zkmist prove ...`.
     if std::env::var("ZKMIST_GEN_VERIFIER").as_deref() == Ok("1") {
-        let sol_code = gen_evm_verifier_sol_code::<AxiomClaimMarker, SHPLONK>(&params, &vk, vec![4]);
+        let sol_code =
+            gen_evm_verifier_sol_code::<AxiomClaimMarker, SHPLONK>(&params, &vk, vec![4]);
         let path = std::path::Path::new("../contracts/src/Halo2Verifier.axiom.sol");
         std::fs::write(path, sol_code).expect("Failed to write verifier");
         eprintln!("      [axiom] verifier regenerated from VK (ZKMIST_GEN_VERIFIER sanity check)");
